+++ ft_strnew_three.c
@@ -6,7 +6,7 @@
 /*   By: cfatrane <cfatrane@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/01/20 11:50:06 by cfatrane          #+#    #+#             */
-/*   Updated: 2017/01/27 20:13:15 by cfatrane         ###   ########.fr       */
+/*   Updated: 2017/01/27 20:28:30 by cfatrane         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	***ft_strnew_three(size_t x, size_t y, size_t z)
 {
-	int		i;
-	int		j;
+	size_t	i;
+	size_t	j;
 	char	***tab;
 
 	i = -1;
