+++ ft_tabnew_three.c
@@ -6,17 +6,17 @@
 /*   By: cfatrane <cfatrane@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/01/21 14:24:48 by cfatrane          #+#    #+#             */
-/*   Updated: 2017/01/21 14:24:55 by cfatrane         ###   ########.fr       */
+/*   Updated: 2017/01/28 11:54:20 by cfatrane         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	***ft_tabnew_three(int x, int y, int z)
+int	***ft_tabnew_three(size_t x, size_t y, size_t z)
 {
-	int	i;
-	int j;
-	int	***tab;
+	size_t	i;
+	size_t	j;
+	int		***tab;
 
 	i = -1;
 	if (!(tab = (int***)malloc(sizeof(*tab) * y)))
